Styling and component changes and refactoring
