+++ src/components/search/SearchForm.js
@@ -5,10 +5,19 @@ import styled from 'styled-components';
 import { Input } from 'antd'
 
 const Wrapper = styled.div`
-    .ant-btn-primary {
-        background-color: #547862;
-        border-color: #547862;
+    background-color: #f3f6f5;
+    padding: 16px 0;
+    
+    .innerWrapper {
+        width: 90%;
+        margin: auto;
+
+        .ant-btn-primary {
+            background-color: #547862;
+            border-color: #547862;
+        }    
     }
+    
 `;
 
 const SearchForm = props => {
@@ -23,27 +32,28 @@ const SearchForm = props => {
     
     const handleSearch = e => {
         if(e.length){
-            props.getGoogleResults(searchQ.q);
+            props.getGoogleResults(searchQ.q );
         }else{
             setSearchQ({q: ''});
             props.clearSearchResults();
         }
     }
-    // onSearch, onEnter
-    // value => props.getGoogleResults(value)
+
     return (
-        <Wrapper>    
-            <Input.Search name="q"
-                allowClear
-                aria-label="search-box"
-                placeholder="Search for a book" 
-                size="large"
-                onSearch={handleSearch}
-                onEnter={handleSearch}
-                onChange={handleChange}
-                value={searchQ.q}
-                enterButton
-            />
+        <Wrapper>
+            <div className="innerWrapper">
+                <Input.Search name="q"
+                    allowClear
+                    aria-label="search-box"
+                    placeholder="Search for a book" 
+                    size="large"
+                    onSearch={handleSearch}
+                    onEnter={handleSearch}
+                    onChange={handleChange}
+                    value={searchQ.q}
+                    enterButton
+                />
+            </div>
         </Wrapper>
     )
